Added logs, fixed UserManager bug, created settings dir
The logs are very basic and using Django logging. The bug with
UserManager is that it inherits BaseUserManager which does not have
create_superuser method. Therefore, creating superuser for Admin from
command line fails. Inhering from UserManager solves the issue. To
prevent tests from generating logs the settings for dev and test are now
separate.